fix for data type long

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -8,15 +8,15 @@
  */
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int count = 0;
+	unsigned long int count = 0;
 	unsigned int len = 0;
-	unsigned int a = 0x01;
+	unsigned long int a = 0x01;
 
 
 	count = n ^ m;
 	while (count != 0)
 	{
-		len = len + (a & count);
+		len = len + (unsigned int)(a & count);
 		count = count >> 1;
 	}
